requirements.txt
pillow==12.0.0
zpl==0.1.13
